Update patch set 5
Patch Set 5:

aosp/1242301 now does some of this in a different way, after experimenting with some options around this. Based on the results I saw on microbenchmarks, avoiding sched_yield doesn't really help. Inflating makes the lock significantly more expensive, so we want to avoid that if we can. If we just spin 50 times without sched_yield, we see a lot more inflation. We do want to spin for a bit at the beginning, which we now do. That seems to somewhat improve H2 performance on some platforms, but it looks inconsistent enough to me that it's actually a big deal.

Abandon this one?

Patch-set: 5
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>