Make deoptimize control flow explicit
Currently deoptimize is one of the only predicated instructions and
has an unusual 'never-return' behavior. This changes the graph to make
the control flow explicit. This only affects the barrier type of
HDeoptimize instructions and does not change the guard-value
deoptimize instructions.

In order to ensure that later optimizations can continue to run
correctly an extra induction-variable-analysis pass and another
bounds-check-elimination pass also needed to be added. These passes
refresh IVA data that new blocks can invalidate and clean-up
unreachable deoptimize blocks that the instruction-simplifier would
have previously dealt with.

Bug: 67037140
Test: ./test.py --host
Change-Id: I407ffc32e2b369bded11fb89b377a52b275fd572