Update patch set 1
Patch Set 1:

(2 comments)

Out of curiosity, does this have any immediate impact on code quality/performance, or is it more an enabler for better stuff to come?

I really like the idea of simplifying the number of control flow cases we have to deal with. It should simplify a lot of the compiler.

Patch-set: 1