Fix deadlock bug in test 2005
Test 2005 checks that redefinition is possible even when some of the
threads involved are suspended. It does this by spawning a bunch of
threads, suspending all of them, redefining a class those threads are
using, and then check that the results of the threads calculations are
consistent. This could cause a problem since (due to the structure of
the code) a test thread could be responsible for initializing the
class being redefined. If this is still ongoing when all threads are
suspended then the main test thread will get deadlocked waiting for
the class to finish initializing (since the class must have either not
started or already finished initialization for redefinition to be
possible).

To fix this issue we just force the class to be initialized before
running the main test code.

Bug: 148341959
Test: ./test.py --host
Change-Id: Ic1bf59ffd760b97b3e173ee48d3ce7ad9bb6c3ce