Update patch set 1
Patch Set 1:

The failures are a clear indication that this is fixing the right problem; we seem to be regularly timing out trying to suspend threads. Trivial to fix, but it's probably worth trying to get the code to crash in that scenario to see if we can eliminate the root cause before submitting this.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0