Move the logging line interleave lock from libbase to ART
libbase will now print up to 4k chunks of data to logd at once and
it will print the full log message to stderr at once, so there is no
need for a majority of users to hold a lock during logging to prevent
other threads from interleaving log messages between lines of an
individual log message.

ART is historically the only user of libbase that strongly cared to
have this lock to prevent the interleaving in all cases, thus this
lock is moved to ART only.

Bug: 65062446
Bug: 153824050
Test: logging unit tests
Change-Id: I4b9e13fcc8996c419cb7472f545589f860a3e510