Update patch set 1
Patch Set 1:

Thanks for fixing this, and sorry for breaking it! I'll try to remember to run the RI tests before landing in the future.

Do we run RI tests as part of presubmit? That'd let us prevent these kinds of issues in the future.

Patch-set: 1
Reviewer: Gerrit User 1335232 <1335232@85c56323-6fa9-3386-8a01-6480fb634889>