Update patch set 1
Patch Set 1: Code-Review+2

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > > Patch Set 1:
> > > > > 
> > > > > It seems little hacky.
> > > > > Alternatively, may I propose passing the right flags to the art script instead (as we do for x86)?
> > > > > CL uploaded... I have not tested if it works.
> > > > 
> > > > David, is this CL of yours already uploaded? I did a quick search but could not find it.
> > > 
> > > Found it; sorry, I was looking in the wrong Gerrit instance…
> > 
> > If David's change fixes the issue, I think it would be a better solution.
> 
> It doesn't fix the issue with Compile benchmarks, which invoke dex2oat from /system/bin directly. We could fix that on golem as well. Do you prefer that?

Oh right, we don't generate that symlink for Golem builds because we don't build that part of the platform.

I think we can use you change as a stop-gap measure, but shouldn't we disable ARM Compile benchmarks for ARM altogether? With the current set-up, those benchmarks would run the same `dex2oat` binary on both ARM and ARM64 targets, which is a bit of a waste of resources…

Patch-set: 1
Reviewer: Gerrit User 1052644 <1052644@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2