Relax 2033-shutdown-mechanics result comparison
It appears the daemon thread often ends up hanging early during
shutdown, truncating the output. We actually want daemon threaads
to hang at some point during this process, so it's not clear this
is bad. Consider any prefix of the expected output to be correct.

It would still be nice to better understand what's going on here.

Bug: 155098113
Test: testrunner.py --host -b 2033-shutdown-mechanics
Change-Id: Ib0f510af7cf69202ab8151d637a41be2f6869ff2