Finish move of jar compilation rules from makefile to soong.
The rules have already been previously ported to soong, but they were
only used for atest.  Always use them to simplify the makefile.

This makes the gtest modules in .bp files depend on the needed jars,
which ensures that they will be copied next to the test binary.
This is needed as atest can not access them in the current location.

It also simplifies the tradefed xml since the manual copying
can be removed for the apex-based device testing.

Test: test-art-host-gtest
Bug: 147819342
Change-Id: I54d92eca88fc04c949209d490e838d0a92ce8f87