Code sinking can move around LoadString that can throw.
The test accidentally used a string part of the boot image, which means
we know the instruction won't throw. However, a change in the boot
classpath meant the string "a" was not part of the boot image anymore,
and the test started failing.

The CL now handles the case the LoadString might throw, and treat it
like NewInstance/NewArray.

Test: 672-checker-throw-method
Bug: 156559242
Change-Id: If9df2ed2c7c39c56254970172e315ec5113db64e