Update patch set 5
Patch Set 5:

I've added a comment on go/art-partial-lse-doc, but I'll copy it here for reference:

So I suggest limiting the changes to the LSE pass: treat HDeoptimize environment inputs as non-escaping there, and when removing the allocations, find the places where they flow in a HDeoptimize. If one is found, then at the end of LSE, you adjust the control flow to create a block which will do the materialization of the object (as you intended).

Patch-set: 5
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>