Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The intent is to make eg InstructionSimplifier have two modes: one where it's not loop friendly and one where it is?

Yes, you are right. It will be used as:

void InstructionSimplifierVisitor::VisitRem(HRem* rem) {
  if (rem->IsInLoop() && be_loop_friendly_) {
    return;
  }
  SimplifyRemWithConstantDivisor(rem);
}

> 
> My understanding is that we should do loop analysis (and induction variable analysis) just before doing loop-oriented optimizations instead.

The current sequence is: licm+induction_var_analysis+BCE+loop_optimizations.
InstructionSimplifier happens several times before the sequence. It was easy to add support of recognition of an optimized HRem to BCE. However it is not so easy with loop stuff.

Patch-set: 1