Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > The intent is to make eg InstructionSimplifier have two modes: one where it's not loop friendly and one where it is?
> > 
> > Yes, you are right. It will be used as:
> > 
> > void InstructionSimplifierVisitor::VisitRem(HRem* rem) {
> >   if (rem->IsInLoop() && be_loop_friendly_) {
> >     return;
> >   }
> >   SimplifyRemWithConstantDivisor(rem);
> > }
> > 
> > > 
> > > My understanding is that we should do loop analysis (and induction variable analysis) just before doing loop-oriented optimizations instead.
> > 
> > The current sequence is: licm+induction_var_analysis+BCE+loop_optimizations.
> > InstructionSimplifier happens several times before the sequence. It was easy to add support of recognition of an optimized HRem to BCE. However it is not so easy with loop stuff.
> 
> I'm not following - if all instruction simplifiers run before the various analysis, then you shouldn't need this change?

Let's look at an example, 530-checker-loops4:

  public static int geo4(int a) {
    for (int i = 0; i < 10; i++) {
      a %= 7;  // a wrap-around induction
    }
    return a;
  }

It is expected that the loop is completely removed. The returned result is the calculated last value.
HLoopOptimization and InductionVarRange take HRem and calculate the last value. 
Instruction simplifiers running before the induction_var_analysis+loop_optimizations transform 'a %= 7' into a sequence of instructions: Div, Shl, Sub, Add. So when HLoopOptimization and InductionVarRange look at them they cannot calculate the last value because they don't recognize that this is actually HRem. Implementing such recognition in HLoopOptimization/InductionVarRange will complicate their code. This recognition is actually to undo the HRem optimization, to do analysis, to do loop optimizations and to optimize HRem again. 
So such HRem instruction simplification is not loop friendly and should be done after loop optimizations. It is Okey to do the simplification if HRem is not in a loop.

Hope this makes things clearer.

Patch-set: 1