Update patch set 2
Patch Set 2:

> patch2 works fine with Android R.

Ah, I see where the confusion comes from. The format of boot jar list (as dumped by art/tools/bootjars.sh) has changed between R and AOSP. In R it is a list of names, and in AOSP it is a list of colon-separated apex-name pairs. The script needs to handle both cases (your PS2 works in R but not AOSP, and my code in the comment to PS1 works in AOSP but not in R). I will post an update shortly, after I test it in both AOSP and R.

> I am not sure that if the Usage2 is well liked.

I certainly like the idea to have a dump-only mode (every time I have to patch the script, I end up inserting  temporary `echo` commands). But please do separate it in a follow-up patch, as it is not related to the current fix.

Patch-set: 2