Update patch set 1
Patch Set 1:

Thanks Martin

We should discuss this. It's be good to check with the soong folk whether the libnativehelper change is outside the expected operating range, whether we can just fix this in soong, or whether we can avoid using sdk_version libnativehelper (e.g. re-write the tests to be gtests).

Patch-set: 1