Enable TARGET_BUILD_UNBUNDLED in master-art (reland).
This is the way to disable assumptions in the build system that a full
system image is to be built.

Soong internally checks for the absence of frameworks/base to detect a
reduced master-art tree and disable some dependencies. That check is at
odds with TARGET_BUILD_UNBUNDLED=true, so for now let's enable
TARGET_BUILD_UNBUNDLED only on the same condition.

This relands https://r.android.com/1337182 with changes in two
testrunner scripts that were missed.

Test: art/tools/buildbot-build.sh on AOSP master and master-art
Test: m installclean && art/build/apex/runtests.sh
  on AOSP master and master-art
Test: rm -rf out && \
      art/test/testrunner/run_build_test_target.py art-linux-bionic-x64
  on AOSP master and master-art
Test: art/test/testrunner/run_build_test_target.py art-test
  on AOSP master and master-art
Test: art/test/testrunner/testrunner.py --build-target
  on AOSP master and master-art
Test: art/test/testrunner/testrunner.py --target --64 --optimizing && \
      art/tools/run-gtests.sh
  in test chroot on device
  on AOSP master and master-art
Test: art/tools/golem/build-target.sh \
      --machine-type=android-armv8 --golem=art-jit
  on AOSP master and master-art
Bug: 159109002
Bug: 157549171
Change-Id: Ie98af1098ea1132d3db7ed0a8bf47b54e0848e57