[TEMP] Estimate quickening rewrite impact on interpreter cache.
Record statistics for InterpreterCache usage that could be
avoided with new bytecodes for field and method references
to the defining class. Print these statistics to logcat when
the zygote or system server profile compilation finishes.
(To be used on jitzygote configuration on x86 as the x86-64
nterp and the arm/arm64 mterp access the cache directly from
assembly, skipping the collection of some stats.)

On aosp_cf_x86_phone-userdebug/jitzygote, implementing new
bytecodes for iget*/iput*, sget*/sput*, invoke-static*,
invoke-direct* and invoke-virtual* could avoid around 50% of
lookups (`avoidable_stores/misses`: zygote ~53%, system
server ~48%). Individual instruction groups contribute:
                  zygote   system server
  iget*/iput*      ~36%        ~35%
  sget*/sput*       <1%         ~2%
  invoke-static*    ~2%         ~3%
  invoke-direct*    ~8%         ~4%
  invoke-virtual*   ~6%         ~4%
Also important is the `avoidable/(misses-avoidable_stores)`
that shows the proportion of non-replacable bytecodes that
would yield a cache hit instead of a miss because an entry
would not have been evicted by a replacable bytecode; in
the case of all bytecodes being implemented it shows ~23%
for zygote and ~10% for system server.

Test: n/a
Change-Id: Ia2ee4a36d041f016a80ca97a66d2fad6464e76bc