Update patch set 1
Patch Set 1:

Thanks Vladimir for the numbers.  For the record my plan is:

1) Remove use of current quickening instructions, do my best to mitigate impact, and measure the performance regression.

2) If the regression is not negligible, re-purpose the existing quickening bytecodes as the class-local index variants, and measure performance gain.

We can only remove quickening if both changes together have acceptable performance regression.
I also considering re-purposing only the most common byte-codes to free a bit of the opcode space.

I am reluctant to jump directly to 2) because there slim chance we might get away with the default opcodes, freeing the quicken opcode space altogether, and not extending the spec.  Also since we are only going to handle class-local indices, the mitigations from 1) might be essential.

Patch-set: 1