Update patch set 32
Patch Set 32:

(3 comments)

Hi Neeraj,

Thanks for another PS in the series.

A few comments on the tests (as I commented before I would prefer not to go into reviewing the implementation (.cc) before the tests are fully provided and reviewed):

1. It is not clear for which types/categories of loops the optimization will be applied and for which it won't.
  - Induction kind (kInvariant, kLinear, kPolynomial, kGeometric, kWrapAround, kPeriodic) 
  - Number of basic blocks (more than two?).
  - while loops (if the last instruction of the loop header is not HIf exit).
Basically if "This optimization is applied to all loops except for A, B, and C types" then I would suggest there should be negative tests for A, B, and C. E.g. when in the implementation in the corresponding method there is "return false" so the optimization is not applied then it might be worth to introduce a test to cover that case.

2. I am sorry, I didn't get are there still any cases when an extra induction is introduced now? I'd suggest you could put a detailed comment around this into the sources/tests rather than having in only as a gerrit comment here.

Thanks,
Artem.

Patch-set: 32
Attention: {"person_ident":"Gerrit User 1367653 \u003c1367653@85c56323-6fa9-3386-8a01-6480fb634889\u003e","operation":"ADD","reason":"reviewer or cc replied"}