WIP: Partial LSE analysis & store removal
This is the first piece of partial LSE for art. This CL adds analysis
tools needed to implement partial LSE and improves LSE so that it will
remove stores that are provably observable based on the location
they occur. For example:

```
Foo o = new Foo();
if (xyz) {
  check(foo);
} else {
  foo.x = 12;
}
```

The store of 12 can be removed because the only escape in this method
is unreachable and was not executed by the point we reach the store.

The actual impact of this change is rather minor. In the (device) BCP
we are only able to remove ~80 stores. Across the rest of the AOSP java
files there are 280 stores removed when everything is compiled with
'speed'.

Test: art_compiler_tests
Test: treehugger
Bug: TBD
Change-Id: I0edc8d6b73f7dd329cb1ea7923080a0abe913ea6