Partial LSE analysis & store removal
This is the first piece of partial LSE for art. This CL adds analysis
tools needed to implement partial LSE and improves LSE so that it will
remove stores that are provably observable based on the location
they occur. For example:

```
Foo o = new Foo();
if (xyz) {
  check(foo);
  foo.x++;
} else {
  foo.x = 12;
}
return foo.x;
```

The store of 12 can be removed because the only escape in this method
is unreachable and was not executed by the point we reach the store.

The main purpose of this CL is to add the analysis tools needed to
implement partial Load-Store elimination. Namely it includes tracking
of which blocks are escaping and the groups of blocks that we cannot
remove allocations from.

The actual impact of this change is incredibly minor, being triggered
only once in a AOSP code. go/lem shows only minor effects to
compile-time and no effect on the compiled code. There is a
non-trivial increase in peak memory use.

Test: art_compiler_tests
Test: treehugger
Bug: 67037140
Bug: 160804802
Change-Id: I0edc8d6b73f7dd329cb1ea7923080a0abe913ea6