Suppress explicit gc logging in libcore tests
These logs are so frequent it is difficult to read the output logs. On
a recent run I got ~86% of the lines in the output were the explicit
GC logs. This bloated the log file to ~80 mb, making examining it with
a web-browser difficult. This change should drastically reduce the
size.

Test: manual
Change-Id: I2cdc77814891d125f6a6bb69decd2c68b3fa0f5b