art: Remove dependencies on platform unavailable modules
The device variants of these tests depend on tools such as
`dex2oatd` which are not available to the platform. The tools
are provided by the ART APEX (com.android.art.debug / .testing).

This also breaks darwin host build coincidentally, because on
linux host build the build system confuses the host and device
variants of `dex2oatd`, so the build system mistakenly thought
device variant of "dex2oatd" exists (while in reality only the
host variant exists.)
On darwin host however, both device and host variants of
`dex2oatd` doesn't exist, thus triggering a build break.

So for the host variants of these tests, the tools
dependencies should be specified as
`host.required: [<tool name>]` and for the device variants
`android.required: ["com.android.art.testing"]`.
However these device testcases are already packaged in the
ART testing APEX, so the device dependency should be omitted.

Bug: 162102724
Bug: 7456955
Test: TH build mac host target
Change-Id: I8f3435ff11f19f7191e87b0728e6d524bbd0f559